+++ src/Makefile.am
@@ -19,8 +19,9 @@
 AM_CFLAGS = $(HDF5_CPPFLAGS) $(GSL_INCS) $(FFTW_INCS) $(NUMA_INCS) \
         $(GRACKLE_INCS)  $(SUNDIALS_INCS) $(CHEALPIX_CFLAGS) $(LUSTREAPI_INCS)
 
-# Add HIP Path
+if HAVEHIP
 AM_CFLAGS += -D__HIP_PLATFORM_AMD__
+endif
 
 # Assign a "safe" version number
 AM_LDFLAGS = $(HDF5_LDFLAGS) $(FFTW_LIBS)
@@ -37,6 +38,8 @@ EXTRA_LIBS = $(GSL_LIBS) $(GMP_LIBS) $(HDF5_LIBS) $(FFTW_LIBS) $(NUMA_LIBS) \
 MPI_LIBS = $(PARMETIS_LIBS) $(METIS_LIBS) $(MPI_THREAD_LIBS)
 MPI_FLAGS = -DWITH_MPI $(PARMETIS_INCS) $(METIS_INCS)
 
+CUDA_FLAGS = $(CUDA_CFLAGS) -DWITH_CUDA
+
 # Build the libswiftsim library and a convenience library just for the gravity tasks
 lib_LTLIBRARIES = libswiftsim.la
 noinst_LTLIBRARIES = libgrav.la
@@ -49,6 +52,7 @@ endif
 # Build a cuda version too?
 if HAVECUDA
 lib_LTLIBRARIES += libswiftsim_cuda.la
+noinst_LIBRARIES = libswift_CUDA.a
 if HAVEMPI
 lib_LTLIBRARIES += libswiftsim_mpicuda.la
 endif
@@ -56,6 +60,7 @@ endif
 
 # Build a hip version too?
 if HAVEHIP
+# TODO: Needs fixing. See what we did for cuda.
 lib_LTLIBRARIES += libswiftsim_hip.la
 if HAVEMPI
 lib_LTLIBRARIES += libswiftsim_mpihip.la
@@ -182,6 +187,7 @@ endif
 
 GPU_SOURCES = 
 if HAVECUDA
+# list here files to be compiled without nvcc, but needed for execution with GPU
 GPU_CUDA_SOURCES = 
 GPU_CUDA_SOURCES += runner_gpu_pack_functions.c
 GPU_SOURCES+=$(GPU_CUDA_SOURCES)
@@ -246,6 +252,17 @@ AM_SOURCES += $(SPHM1RT_RT_SOURCES)
 AM_SOURCES += $(GEAR_RT_SOURCES)
 AM_SOURCES += $(GPU_SOURCES)
 AM_SOURCES += swift_lustre_api.c
+
+# sources needed for the cuda build
+CUDA_SOURCES = 
+if HAVECUDA
+CUDA_SOURCES += runner_main_clean.cu
+CUDA_SOURCES += cuda/GPU_data_buffers.cu  cuda/GPU_runner_functions.cu 
+CUDA_SOURCES += runner_doiact_functions_hydro_gpu.h
+endif
+
+# common sources needed for any GPU build
+GPU_COMMON_SOURCES = runner_gpu_pack_functions.c
 
 # Include files for distribution, not installation.
 nobase_noinst_HEADERS = align.h approx_math.h atomic.h barrier.h cycle.h error.h inline.h kernel_hydro.h kernel_gravity.h runner_gpu_pack_functions.h
@@ -546,6 +563,11 @@ nobase_noinst_HEADERS += fvpm_geometry.h fvpm_geometry_struct.h
 nobase_noinst_HEADERS += fvpm_geometry/None/fvpm_geometry.h fvpm_geometry/None/fvpm_geometry_struct.h
 nobase_noinst_HEADERS += fvpm_geometry/Gizmo/fvpm_geometry.h fvpm_geometry/Gizmo/fvpm_geometry_struct.h
 nobase_noinst_HEADERS += fvpm_geometry/Gizmo/MFM/fvpm_geometry.h fvpm_geometry/Gizmo/MFV/fvpm_geometry.h
+nobase_noinst_HEADERS += gpu_pack_vars.h
+
+nobase_noinst_HEADERS += cuda/BLOCK_SIZE.h  cuda/device_functions.h  #cuda/GPU_data_buffers.h  
+nobase_noinst_HEADERS += cuda/GPU_runner_functions.h  cuda/part_gpu.h
+nobase_noinst_HEADERS += gpu_pack_vars.h runner_gpu_pack_functions.h
 
 # Sources and special flags for the gravity library
 libgrav_la_SOURCES = runner_doiact_grav.c
@@ -572,29 +594,29 @@ libswiftsim_mpi_la_LIBADD = $(GRACKLE_LIBS) $(VELOCIRAPTOR_LIBS) $(MPI_LIBS) lib
 
 # Sources and flags for regular CUDA library
 libswiftsim_cuda_la_SOURCES = $(AM_SOURCES)
-libswiftsim_cuda_la_CFLAGS = $(AM_CFLAGS) $(CUDA_CFLAGS) -DWITH_CUDA
-libswiftsim_cuda_la_CXXFLAGS = $(AM_CFLAGS) $(CUDA_CFLAGS) -DWITH_CUDA
-libswiftsim_cuda_la_LDFLAGS = $(AM_LDFLAGS) $(EXTRA_LIBS) $(CUDA_LIBS)
+libswiftsim_cuda_la_CFLAGS = $(AM_CFLAGS) $(CUDA_FLAGS)
+libswiftsim_cuda_la_LDFLAGS = $(AM_LDFLAGS) $(EXTRA_LIBS) $(CUDA_LDFLAGS) $(CUDA_LIBS)
 libswiftsim_cuda_la_SHORTNAME = cuda
 libswiftsim_cuda_la_LIBADD = $(GRACKLE_LIBS) $(VELOCIRAPTOR_LIBS) $(MPI_LIBS) libgrav.la
 
 # Sources and flags for regular HIP library
+# TODO: Needs fixing. See what we did for CUDA.
 libswiftsim_hip_la_SOURCES = $(AM_SOURCES)
 libswiftsim_hip_la_CFLAGS = $(AM_CFLAGS) $(HIP_CFLAGS) -DWITH_HIP
-libswiftsim_hip_la_CXXFLAGS = $(AM_CFLAGS) $(HIP_CFLAGS) -DWITH_HIP
+# libswiftsim_hip_la_CXXFLAGS = $(AM_CXXLAGS) $(HIP_CXXLAGS) -DWITH_HIP
 libswiftsim_hip_la_LDFLAGS = $(AM_LDFLAGS) $(EXTRA_LIBS) $(HIP_LIBS) -lamdhip64
 libswiftsim_hip_la_SHORTNAME = hip
 libswiftsim_hip_la_LIBADD = $(GRACKLE_LIBS) $(VELOCIRAPTOR_LIBS) $(MPI_LIBS) libgrav.la
 
 # Sources and flags for MPI CUDA library
 libswiftsim_mpicuda_la_SOURCES = $(AM_SOURCES)
-libswiftsim_mpicuda_la_CFLAGS = $(AM_CFLAGS) $(MPI_FLAGS) $(CUDA_CFLAGS) -DWITH_CUDA
-libswiftsim_mpicuda_la_CXXFLAGS = $(AM_CFLAGS) $(MPI_FLAGS) $(CUDA_CFLAGS) -DWITH_CUDA
-libswiftsim_mpicuda_la_LDFLAGS = $(AM_LDFLAGS) $(MPI_LIBS) $(EXTRA_LIBS) $(CUDA_LIBS)
+libswiftsim_mpicuda_la_CFLAGS = $(AM_CFLAGS) $(MPI_FLAGS) $(CUDA_FLAGS)
+libswiftsim_mpicuda_la_LDFLAGS = $(AM_LDFLAGS) $(MPI_LIBS) $(EXTRA_LIBS) $(CUDA_LDFLAGS) $(CUDA_LIBS)
 libswiftsim_mpicuda_la_SHORTNAME = mpicuda
 libswiftsim_mpicuda_la_LIBADD = $(GRACKLE_LIBS) $(VELOCIRAPTOR_LIBS) $(MPI_LIBS) libgrav_mpi.la
 
 # Sources and flags for MPI HIP library
+# TODO: Needs fixing. See what we did for CUDA.
 libswiftsim_mpihip_la_SOURCES = $(AM_SOURCES)
 libswiftsim_mpihip_la_CFLAGS = $(AM_CFLAGS) $(MPI_FLAGS) $(HIP_CFLAGS) -DWITH_HIP
 libswiftsim_mpihip_la_CXXFLAGS = $(AM_CFLAGS) $(MPI_FLAGS) $(HIP_CFLAGS) -DWITH_HIP
@@ -602,10 +624,26 @@ libswiftsim_mpihip_la_LDFLAGS = $(AM_LDFLAGS) $(MPI_LIBS) $(EXTRA_LIBS) $(HIP_LI
 libswiftsim_mpihip_la_SHORTNAME = mpihip
 libswiftsim_mpihip_la_LIBADD = $(GRACKLE_LIBS) $(VELOCIRAPTOR_LIBS) $(MPI_LIBS) libgrav_mpi.la
 
+if HAVECUDA
+CUDA_OBJECTS = runner_main_clean.o
+CUDA_OBJECTS += cuda/GPU_data_buffers.o  cuda/GPU_runner_functions.o
+CUDA_DLINK_OBJECTS = runner_main_clean.dlink_o
+CUDA_DLINK_OBJECTS += cuda/GPU_data_buffers.dlink_o cuda/GPU_runner_functions.dlink_o
 
-#subdir
-SUBDIRS = . cuda
-SUBDIRS += . hip
+libswift_CUDA_a_SOURCES = $(CUDA_SOURCES) $(GPU_COMMON_SOURCES)
+# libswift_CUDA_a_LIBADD = $(CUDA_OBJECTS) $(CUDA_DLINK_OBJECTS)
+libswift_CUDA_a_LIBADD = $(CUDA_DLINK_OBJECTS)
+
+AM_NVCCFLAGS = -I$(top_srcdir) -I$(top_srcdir)/src
+# wildcard making recipes for *.cu -> *.o
+.cu.o: 
+	$(NVCC) $(NVCC_FLAGS) $(AM_NVCCFLAGS) $(AM_CFLAGS) $(CUDA_FLAGS) $< -dc -o $@
+
+# wildcard making recipes for *.cuo -> *.dlink_o
+.o.dlink_o: 
+	$(NVCC) $(NVCC_FLAGS) $(AM_NVCCFLAGS) $(AM_CFLAGS) $(CUDA_FLAGS) $< -dlink -o $@
+endif
+
 
 # Versioning. If any sources change then update the version_string.h file with
 # the current git revision and package version.
