+++ src/clocks.h
@@ -20,11 +20,7 @@
 #define SWIFT_CLOCKS_H
 
 /* Config parameters. */
-#ifdef WITH_CUDA
-#include "../config.h"
-#else
 #include <config.h>
-#endif
 /* System includes. */
 #include <sys/times.h>
