+++ src/cuda/GPU_runner_functions.h
@@ -5,7 +5,10 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
+#include <cuda_runtime.h>
+
 #include "part_gpu.h"
+
 void launch_density_kernel(struct part_soa parts_soa, int *d_task_first_part,
                            int *d_task_last_part, float d_a, float d_H,
                            const char *loop_type, cudaStream_t stream,
