+++ src/cuda/device_functions.h
@@ -1,6 +1,11 @@
 #ifndef DEVICE_FUNCTIONS_H
 #define DEVICE_FUNCTIONS_H
 #include "../../config.h"
+#include <math.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
 
 /* Local headers. */
 // #include "../dimension.h"
@@ -34,9 +39,8 @@
  * Alpha can be set in the parameter file.
  * Beta is defined as in e.g. Price (2010) Eqn (103) */
 #define const_viscosity_beta 3.0f
-#ifdef WITH_CUDA
-extern "C" {
-#endif
+
+
 /**
  * @brief Returns the argument to the power given by the dimension plus one
  *
@@ -142,8 +146,10 @@ __device__ void d_kernel_deval(float u, float *__restrict__ W,
   *dW_dx = dw_dx * kernel_constant * kernel_gamma_inv_dim_plus_one;
 }
 
-#ifdef WITH_CUDA
+#ifdef __cplusplus
 }
 #endif
+
+
 
 #endif  // DEVICE_FUNCTIONS_H
